hms_gitgang/requirements.txt
amqp==5.2.0
ansicon==1.89.0
arrow==1.3.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.0
blessed==1.20.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
decorator==4.4.2
Django==5.0.8
django-allauth==64.1.0
django-cors-headers==4.4.0
django-oauth-toolkit==2.4.0
django-picklefield==3.2
django-sslserver==0.22
djangorestframework==3.15.2
ffmpeg-python==0.2.0
future==1.0.0
hyperlink==21.0.0
idna==3.7
imageio==2.35.0
imageio-ffmpeg==0.5.1
incremental==24.7.2
jinxed==1.3.0
jwcrypto==1.5.6
kombu==5.4.0
m3u8==6.0.0
moviepy==1.0.3
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.84
pillow==10.4.0
proglog==0.1.10
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==3.5.3
requests==2.32.3
service-identity==24.1.0
setuptools==72.2.0
six==1.16.0
sqlparse==0.5.1
tqdm==4.66.5
Twisted==24.7.0
txaio==23.1.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.0.2
